clicommands: typedef setup command should try to create the full path where the typedefs are stored (#397)
I ran into a small bug with this script where I had `~/.lute` on my
device, but not the rest of the path. The setup script only checks if
this directory exists and assumes that the rest of the path is there if
`~/.lute` is there. This can cause errors with writing the type def
files because the subpaths won't exist.

This change also refactors the code that makes the subdirs by letting
you pass the full path and then making the subdirectories as you go.