infra: make some minor improvements to the bootstrap script and the build system (#396)
This PR is a bit of a grab bag of small fixes to both the build system
and the bootstrap script. The important fix here is that we force
`CURL_ZSTD` off so that we won't have linking failures if you _happen_
to have zstd present on your system for other reasons. Otherwise, we
tweak the message a bit when printing the Lute version in CMake, and
tweak the bootstrap script to echo the key commands it's running to make
the whole output from it easier to interpret. We also change the default
behavior for the bootstrap script to assume something vaguely POSIX-y
without specialized needs, which makes the script work on additional
Unix platforms like FreeBSD. Finally, we change `bootstrapped-lute` to
`lute0` for the initial version of lute generated without the standard
library embedded in it because the term `bootstrapped` here feels
confusing as it reads like it's the _final_ version, rather than an
intermediary. `lute0` is "the first build" and, god help us, if our
process is ever multistaged for bootstrapping, it has an obvious
extension to `lute1`, `lute2`, etc. This may be reasonable to do later
for validation if only to make sure that `lute1` and `lute2` are the
same executable, byte for byte.