Prevent merging failing PRs by making CI aggregator fail when `build`, `test`, or `check-format` jobs fail (#401)
### Problem

Our CI aggregator job was unconditionally running and passing. This
wasn't too much of an issue because contributors weren't merging their
PRs until all checks passed, but PRs set to auto-merge were being merged
automatically.

Because of this, some PRs introduced bugs that got merged in: #386 and
#396.

### Solution

The aggregator job's now depends on the success of the `build`, `test`,
and `check-format` jobs. If any job fails, the aggregator fails, which
will block PRs from being merged.

The bug in #396 was fixed in #400.

To fix the bug introduced in #386, this PR also reverts that PR's
changes to our GitHub Actions setup (using the bootstrap script to
create a `lute` binary). We already have Foreman installed on these
runners and there's no need to use bootstrapping here—in fact, it makes
each CI job quite a bit slower since two fresh builds are needed in
order to run anything.