Enable alias discovery even when ancestors in VFS are technically ambiguous (#405)
Resolves #332.

If navigating up a virtual filesystem yields `Ambiguous` during the
alias discovery phase, the Luau.Require library currently fails
silently. This is an issue, but Lute could also do better here. In the
following filesystem, `main.luau`'s parent is _technically_ ambiguous,
but because filesystems are trees, it's not actually ambiguous whether
we should navigate to the file `folder.luau` or the directory `folder`
(the latter is correct):
- folder.luau
- folder
  - main.luau

Solution: in the case of parent operations in the VFS, ambiguity is
overridden to be treated as a success.

I'll also be putting up a separate PR to fix the silent error in the
Luau.Require library.