[net.serve] Integrate app with main libuv event loop to prevent blocking the task scheduling system (#404)
Resolves #384.

### Problem

The current `net.serve` logic is to schedule a continuation that runs
every task scheduler step, which (1) explicitly calls `run()` on our
`uWS::TemplatedApp<T>`, and (2) schedules itself to run the next frame.
This would work if `run()` were actually `step()`, but uWS doesn't
provide an easy way to step an app.

### Solution

After some investigation, because uWS is configured to use libuv under
the hood, we can patch in the libuv event loop that our task system is
managing by calling `uWS::Loop::get(uv_default_loop())` before creating
the app object. We get many things for free here (e.g. reference
counting for keeping the server alive, managed by libuv), and it's
simpler to reason about. This also means we don't need to explicitly
"run" the app; it simply gets registered to run on the same event loop.

To ensure this works, some minor changes were made to the task scheduler
logic to ensure that we are consuming events from the libuv event loop
if it is live and to ensure we aren't blocking on I/O.