Fix release workflow by stripping away extraneous information from `get_version.cmake` output (#419)
Our release workflow depends on the output of `get_version.cmake` being
the raw version. We added some extra text in #396 that is currently
causing errors during the release tag generation step, so this PR
reverts that change.

Someone reached out about recent changes not being "released" yet, which
was surprising because we're on a nightly release schedule—this was the
root cause.