stdlib: add `system` and boolean environment checks (#416)
### Problem

There was no std lib for `system` related functionalities. 

### Solution

We introduce a `std` library analogue for `system`, where we can now
check the environment with boolean flags.
Closes #167

To test (with huge help from @vrn-sn), we wanted to write a C++ file
that compares the "os" captured by the C++ executable and compares to
the `@std/system` OS. To write that (and other similar output comparison
test cases), we exposed the `CliRuntimeFixture` and a `runCode()`
function that was used by `require.test.cpp` before but we added a
`capture()` function based on the `Luau/ReplFixture` class.

Then, we changed some function signatures to ensure the program args
(`argc` and `argv`) are not statically passed to avoid some tests
failing.

### Example:

```lua
local system = require("@std/system")

if system.linux then
    print("Running on Linux")
end
```

See `examples/system_environment_check.luau` for example use case