fix: cleanup/free filesystem requests and data (#425)
Some filesystem operations didn't clean up and free all their data. This
PR attempts to fix this.

Some general observations:

- `uv_fs_req_cleanup` has to be called on all requests (sync/async). The
documentation doesn't really mention this in much detail. For some
operations, this might be a noop, but that's not part of the API and
libuv is free to change this afaik.
- When handling errors, the request wasn't always free'd.
- The data associated with requests wasn't always free'd.
- Some code was unreachable due to a Lua error thrown before.

These errors seem hard to detect statically. Maybe there could be some
RAII type for typed requests?