stdlib: Update the testing api with some more test lifecycle methods + testing assertions (#411)
This PR comes with some changes to the testing api to reduce the amount
of global state we depend on - when making subcases, we now pass in the
suite object on which the user can operate. Additionally, assertions are
passed in with the test case itself, making them more accessible and
reducing an additional line of code needed to start writing tests.

- [x] {before,after}x{all,each} lifecycle methods
- [x] Table assertions 

For example: 
<img width="463" height="79" alt="image"
src="https://github.com/user-attachments/assets/31003849-ae2c-469d-a23b-415c4135865b"
/>