feat: add `process.execpath` to get the current lute executable (#431)
I was trying to get the code generation to run through CMake (which
would run lute in the build). When integrating this into luthier, I
needed a way to identify the current lute executable to pass to CMake.

This adds `execpath` to `@lute/process`, which mirrors Node.js'
[process.execPath](https://nodejs.org/docs/latest/api/process.html#processexecpath)
to get the lute executable path - this is immune to `exec -a foo lute`
having `foo` as `argv[0]`.


[`uv_setup_args`](https://docs.libuv.org/en/v1.x/misc.html#c.uv_setup_args)
can allocate on some platforms. That memory should be free'd with
[`uv_library_shutdown`](https://docs.libuv.org/en/v1.x/misc.html#c.uv_library_shutdown).
I'm not sure where it's appropriate to call this.

Passing around the `argv0` also seems a bit wrong. Maybe this path
shouldn't be stored in `Runtime`?