Refactor `lute setup` to write to luaurc, add std type defs (#447)
Refactors the `lute setup` command to:
1. Put the regular lute typedefs into a `lute` subdirectory, and add a
`std` subdirectory with the standard library files inside
2. Accept a flag to create a new luaurc, or overwrite the aliases of one
if it exists, to add the type definition folders. Unfortunate caveat is
that the std lib json encoder is not deterministic, so each time it
moves the json keys around - this is pretty annoying and creates a lot
of diff noise. Not sure if there's much to be done about this

Luthier:
1. Fix a backslash escaping bug for std lib and command C++ file gen
that cropped up when I added this
2. Change definition generation to use above structure
3. Use projectRelative more
4. Change order of `generateFilesIfNeeded` calls to fix a bug - CLI
commands rely on the generated type def output, so they should be
generated after type defs

Also fix the trailing comma in the project luaurc which is typically not
valid json

Closes #424

---------

Co-authored-by: ariel <aweiss@hey.com>