std/json: distinguish more explicitly between array and object (#468)
We got some useful feedback from @vocksel about not being able to get a
clean refinement between array and object with the existing types setup,
and the correct solution to this problem here is for us to leverage
another newproxy as a key to distinguish between objects and arrays.
We've provided some additional functions to the interface of `json` both
to construct `object`s and to test against values being `object`s or
`array`s. There's also some clean up in here for json as a library
generally, including correcting its casing now that it's a standard
library and including `object` and `array` as part of the exported types
as well.