stdlib: add `std/fs` and `std/process` to handle file paths as luau `path` and `pathlike` types instead of `string` (#462)
### Problem
Some Lute `fs` and `process` functions have been processing file paths
as `string` types while other places in Lute have been using `path` or
`pathlike` for file path handling/processing. We want to add support for
`path` and `pathlike` file paths to `fs` and `process` functions.

### Solution

We're introducing `fs` and `process` to the Lute standard library as
`@std/fs` and `@std/process` that act as wrappers around the existing
Lute `fs` and `process` files where we can pass in `pathlike` type paths
in `@std/fs` and return `path` type paths in `@std/process` instead of
needing to parse and format each call.

There's also a refactor of the `path/` directory that separates out type
definitions into a separate file outside of the `init`. This was
partially due to a previous attempt of handling path types and running
into cyclic dependencies due to the nature of the `process.luau` file,
but seemed reasonable to keep around for potential future use.

Added tests and found some minor bugs in `process.cpp` that is also
fixed in this pr