feature: Implement a Static Require Tracer in order to bundle files correctly when bytecode compiling them. (#452)
This PR implement a static require tracer for Luau as a prerequisite to
enabling multi file bytecode compilation. This module statically
resolves requires relative to a given entry module, effectively
snapshotting the state of the filesystem for embedding in a lute
executable. This module graph will be embedded in the lute executable
when `lute compile` is invoked and will let the bundle virtual file
system [PR](https://github.com/luau-lang/lute/pull/450) correctly
navigate the directory structure to resolve `require` calls and match up
those paths to the pre-compiled bytecode embedded in the bundle.