Fix multiple runtime APIs that were incorrectly using `std::string::reserve` (#478)
### Problem

Multiple runtime APIs incorrectly were using `std::string::reserve`
where they should have been using `std::string::resize`. This led to UB
that seemed to behave fine at runtime, but the introduction of #467
caused all of these APIs to break.

### Solution

We introduce `uvutils::getStringFromUv` as a utility API to help reduce
errors when interacting with libuv. Internally, this handles all the
buffer logic and simply returns a value or an error to the caller.

In fixing this, I also discovered two issues with our CI:
1. One of the checks in `path.posix.test.luau` is flipped but was
mistakenly passing because of the previously broken `process.cwd()` API.
2. Our Luau tests don't even run on Windows! 😞 

I've fixed (1) and have made a separate issue for (2): #483.