stdlib: fix memory leak in luau.compile (#485)
Fixed a memory leak which was causing GC for my other changes to fail.
Specifically, there was a `std::string` which wasn't being freed because
it was being moved into a userdata object, so I attached a destructor to
it. I also got rid of the tag which was being applied to the userdata
object being constructed because it wasn't being checked, and we were
already applying a type-specific metatable to it. Thanks @~vrn-sn for
helping me come up the fix!

Also added a missing export for `system.tmpdir`.

---------

Co-authored-by: Varun Saini <61795485+vrn-sn@users.noreply.github.com>