Use `std::destroy_at` instead of manually invoking destructors (#489)
C++17 introduces `std::destroy_at` under the `<memory>` header, which
prevents confusion about which destructor to use when paired with
placement new. For many cases, choosing the correct destructor is
trivial, but for type names that are inside a namespace or aliased, it
can be somewhat confusing which destructor to call. Using
`std::destroy_at` automatically compiles the call to the correct
destructor based on the pointer type.