Use visitor pattern for exhaustivity in IO stream handling (#487)
There are many other places where this pattern would be safer than
`get_if`, but this is the only one where the overload pattern isn't
needed (since all variants are handled identically). I'll need to think
on how we want to introduce the overload pattern (either somewhere
central in Lute or in the Luau repo itself).