feature: Introduce a multifile payload format for embedding into the lute binary (#484)
This PR is an incremental step towards supporting multi-file bytecode
compilation. This PR defines a `LuteExePayload`, for encoding and
decoding the multi file bytecode bundles that will be appended to the
`lute` binary when shipping executables.

Once this PR is in, we can start writing the glue code that will let us
produce a lute executable with multiple .luau files embedded.