refactor: Add some utilities for reading and writing Lute Executables in a more systematic fashion (#490)
This PR adds a `LuteExecutable` wrapper, that handles extracting the
`LuteExePayload` out of a binary and also creating a valid `lute`
executable with the payload injected in correctly. This will set us up
to be able to do cross compilation a bit easier, by separating the
`payload` from the actual act of manipulating an executable file.