stdlib: lute test shouldn't point to xpcall for error msg (#518)
While working on tests for something else, I noticed that in cases where
a test case fails for reasons other than an assertion, the testing
harness reports the source of the error as an `xpcall` call in the
lute's test implementation. After discussing with @~Vighnesh-V, I've
implemented a small workaround where we look lower down in the stack for
the source of the error, so that we can actually point to the test case
causing the failure in failure reporting.