cli: fix the paths used in lute setup to correctly install definitions in the correct location (#530)
When we refactored to using the path library in more places, we broke
`lute setup` and made it dump everything in the local directory instead
of in the correct home directory location. We can fix it fairly easily
by adjusting the paths that get used, but I also just did some better
hygiene on using the path library in setup then. Fixes #517.