Remove infinite loop test for .config.luau (#548)
We already have a test for this in the Luau repository, and it's just
causing issues for luau-lsp here. If there ever is a bug with this
functionality, it needs to be fixed in the Luau repository anyway, so I
think it's okay to remove from here.

I was considering replacing it with a test that created the files in
memory, but it seems like too much hassle for not much benefit.

Resolves https://github.com/JohnnyMorganz/luau-lsp/issues/1246 on our
end.