bugfix: windows path api needs to be able to get the drive letter for a path (#554)
There are a number of tests failing on windows runners because the
github runners execute tests under the D:\ drive. However, many of our
tests assume execution under the C drive. This PR adds a function called
`getdrive(path)` to the win32 path api, as well as fixes some of those
tests to be drive aware.