bugfix: createdirectory creates the wrong path on windows machines (#555)
The `fs.createdirectory` api creates intermediate folders by
incrementally creating each subportion of a path. When the passed in
path is an absolute path, on posix systems we were creating a path that
corresponds to `/`. However, on Windows systems, we were leaving off the
drive letter, which means that these paths were treated as relative.
E.g. instead of creating: `C:\\some\absolute\path`, we were creating
`C:some\absolute\path`, relative to the `cwd`. This caused a number of
file systems tests to fail when creating these directories, because they
were being created in wrong place.