API: luau.typeofmodule v1 (#531)
### Problem

As part of an effort to bring type-directed documentation generation to
Luau, we don't have an automated way of getting the return type of a
Luau module, and this is a broader missing feature as well in Luau for
being able to analyze the exact type of a module.

### Solution

We introduce an API in `@lute/luau` that allows users to get the return
type of a module path by calling `luau.typeofmodule()` and passing in
the module path as a string. We create a `Luau::Frontend` to analyze the
and resolve the Module, including resolving requires to return the full
return type of the specified module! For ease of reviewing, this API
temporarily returns the `string` representation while we bikeshed a Luau
data structure representation of the `Type` that's returned.

As written in `examples/get_module_return_type.luau` an example usage
would be retrieving the return type of `@std/task`, which when we run

```luau
local moduleTypeString = luau.typeofmodule("./lute/std/libs/task.luau")
```

We currently return a string representation of the module type:

```
{
    await: (t: {
        co: thread,
        result: any,
        success: boolean?
    }) -> any,
    awaitall: (...{
        co: thread,
        result: any,
        success: boolean?
    }) -> (...any),
    create: <a..., b...>(f: (a...) -> (b...), ...any) -> {
        co: thread,
        result: any,
        success: boolean?
    }
}
```

Which helps us understand that the `task` library has the following
functions `await`, `awaitall`, and `create` with their respective full
return types.

This is all built on top of @wmccrthy's solid boiler plate of the API
(https://github.com/luau-lang/lute/pull/375) and @vrn-sn 's help with
the resolve requires!!

### Future Work

- We definitely want to return a data structure representation of the
Type instead of the string
- We also want to represent types with aliases instead of returning the
exact type (something like how our `definitions/` files look with the
exported types and functions).
- If the file we're getting the type of ever has errors, then the result
will include the errors because our Frontend mode is set to `NoCheck`,
so there will have to be some work done on resolving the errors in the
module files I think
- Alias resolving isn't fully working yet with `@std/` directly passed
in to the `typemodule` call (see comments below), but will be addressed
shortly