lute parse: fix serialization of union type separators (#563)
This PR fixes a bug where the `|` separator was not being correctly
serialized when it followed an optional option of a union. For example,
the `|` in the union type `number? | string` was not being correctly
serialized, and was instead being consumed as the preceding trivia of
the `string` option. We now _always_ try to consume the separator if
serializing an optional option which isn't the last option of the union.

Added a test checking this functionality.

Addresses #551.

---------

Co-authored-by: ariel <arielweiss@roblox.com>