crypto: implement a secretbox interface for authenticated encryption (#552)
The last remaining bit of work to close out #161 was the desire to have
some kind of API for authenticated encryption (AE) or authenticated
encrypted with additional data (AEAD). I worked a bit on trying to
provide the full AEAD interface from boringssl, but it felt clunky and
bad to use in Luau. I don't want to commit to the runtime providing a
standard interface that is literally reproducing openssl's headers,
that's something that can be done later with FFI. Instead, I want the
crypto library to feel more in the spirit of libsodium, providing
opinionated default ways of solving common cryptography tasks. The
secretbox interface is a well-understood metaphor for AE(AD), and one
that works quite nicely. To mitigate the likelihood of nonce or key
reuse, the interface provided right now seals a message providing the
key and nonce as output. We can extend the API in the future with
optional arguments for nonce and key as needed.