bugfix: lute check supports ~ in luaurc paths (#573)
Users have been having issues with the `.luaurc` generated by the `lute
setup` command. This sets certain aliases up in a path with `~` in them
but the require rfc didn't support navigating to aliases with `~` in
them.

This PR fixes this by allowing the FileVFS to navigate to aliases that
represent paths with `~` in them.

Annoyingly, I had to use the filesystem header to handle the test,
because it sets up a directory under `~` called `lute_test_special` and
writes a file there - if yall have suggestions for how to fix this,
please let me know.