stdlib: Printing AST replacements preserves trivia (#580)
Implemented trivia preservation for AST node replacements by grabbing
the leftmost and rightmost trivia of the node being replaced. Along the
way, I discovered that trivia is often bound to the EOF token, rather
than AST nodes, so I had to make a change to make sure we pass that EOF
token around inside of query. To increase the chances that this token is
present, I want to encourage codemod authors to use the parse function
which returns `ParseResult` (which includes EOF), so I renamed
`parse(...) -> AstStatBlock` to `parseBlock`, and `parseFile(...) ->
ParseResult` to `parse`.

The bulk of this PR is the machinery for getting that trivia, as well as
the tests making sure we correctly reserialize trivia for every AST
node. While writing those tests, I also found a bunch of other bugs.