lute lint: Support multiple lintee files (#617)
Update lute lint to support multiple input files by assuming that all
non option arguments are paths to be linted. If a directory is passed,
we recursively walk the directory for all the files to be linted. We
error handle any files that error while linting so we can keep linting
other files.