tidy: Lute shouldn't pull in the filesystem header (#628)
This [PR](https://github.com/luau-lang/lute/pull/573) introduced a
dependency on `std::filesystem` in the `Lute.Test` target. There are
some known problems with this header, and it was suggested in that PR by
@aatxe that we should manually implement the platform specific file I/O
operations that we need as part of a `FileUtils` library in lute, so
that we can remove our dependency on this header. This PR implements the
necessary operations that we need and removes our dependency on
`std::filesystem`.