Battery for Text Diffing (#620)
Adds a `diffText` battery, that implements a myers-diff esque algorithm
using the notion of an edit graph. An edit graph can be described as:
- src string along x axis
- destination string along y axis
- each edge in the graph represents a diff operation:
	- (n, n) -> (n+1, n) represents a deletion of src[n+1]
	- (n, n) -> (n, n+1) represents an addition of destination[n+1]
	- (n, n) -> (n+1, n+1) exists only if src[n+1] == destination[n+1]
- The shortest path from (0, 0) -> (len(src), len(destination))
represents the shortest-edit-sequence.

Myers-diff works by greedily following diagonal edges (n,n -> n+1,n+1)
but the algorithm is not super intuitive / readable imo.

This implementation stays consistent with the greedy nature but varies
slightly in how the graph is traversed. We use a deque to BFS the edit
graph in a greedy fashion: whenever a node has a diagonal edge, we push
it to the front of the deque, while other edges are always pushed to the
back.

Here is what the output of `examples/diffText.luau` looks like:

<img width="302" height="364" alt="Screenshot 2025-12-04 at 6 02 45 PM"
src="https://github.com/user-attachments/assets/8d2294b8-2cdf-4c2b-b864-d434dfbb6420"
/>


**Follow up work:** support side-by-side diff printing