Adds support for luaurc files in executables from `lute compile` (#626)
The `lute compile` command is built on the `resolveRequire` api. This is
`.luaurc` aware, which means that we can resolve require statements that
have aliases in them.
However, while running the binary that `lute compile` emits, we aren't
able to resolve these aliased requires at runtime, because we don't have
access to the `.luaurc`. This PR adds support for this by:

1. Collecting the paths to .luaurc during static require tracing
2. Injecting this configuration into the final executable.
3. Reading the config back out when running the embedded scripts and
navigating to directories based on the aliases stored.

Currently doesn't support config.luau.

---------

Co-authored-by: ariel <arielweiss@roblox.com>