Surfaces some require errors in `lute run` (#660)
I was running into a scenario where I was trying to `lute run` a file in
a directory containing a folder with the same name as the file (ie
`scratch.luau` and a folder called `scratch`). This was causing an
ambiguous require, but lute would just report that `scratch.luau does
not exist`, which was quite frustrating. I've added some piping to
bubble the error up the stack, plus a test.