Implements `lute pkgrun <file>`: a subcommand that runs Lute with package awareness (#657)
Implement a simple layer that bridges the gap between Loom's generated
`loom.lock` file and Lute's `Package::RequireVfs` (introduced in #339).

Running a file with `lute pkgrun <file>` searches for a lockfile called
`loom.lock` before executing `<file>`. We use regex matching to parse
out the set of dependencies from the lockfile, which are expected to be
installed in a `Packages` folder in the same directory as `loom.lock`.
Using this, we construct a `Package::RequireVfs`, which automatically
injects aliases to dependencies when executing user code.

Because of limitations with the current `loom.lock` format, there are
still several improvements we need to make. I've left a few TODO
comments related to this. Long term, regex matching is also not a great
way to parse `loom.lock`, but I didn't want to embed a C++ TOML parser
or deal with the complexities of embedding `@batteries/toml` and calling
into it from C++. Instead, we can probably use Luau syntax for
`loom.lock` since Lute can easily parse Luau tables. We could
potentially base the format on [`.config.luau`
files](https://rfcs.luau.org/config-luauconfig.html).