Ensures that Lute will not block other runtime threads during basic file system operations (open, close, read, write) (#661)
This PR ensures that `fs.(open, close, read, write)` are non-blocking -
that is, other `Luau` threads that could continue to run, will not
blocked by long i/o. This fixes
https://github.com/luau-lang/lute/issues/642, whereby a very long read,
e.g. to /dev/random could cause the main Lute Runtime thread to block
synchronously (which would prevent stepping the uv event loop as well).
I've tested this locally and it works as expected, but I'm not sure how
to write a test for this in a cross platform way.

This PR also:
- Adds a base `UVRequest` type - responsible for interfacing with the
lute runtime to resume the original luau thread. During destruction,
will invoke the uv_*_cleanup function.
- Makes File Handles a lightuserdata.
- Rewrites `stdfs.writefiletostring` and `stdfs.readstringfromfile` in
terms of the low level `lute` primitives that don't block anymore.
- Removes `readasync`, `writefiletostring` and `readstringfromfile` from
the `lute` api surface.