@std/test Tech Debt: Refactor how we provide debug information via assertions + make runtime error detection less brittle (#651)
This PR attempts to refactor the test library to be slightly more
maintainable. While trying to bring
https://github.com/luau-lang/lute/pull/519 up to date, I experienced a
lot of brittleness with our test libraries ability to detect when a test
failed due to runtime errors. The test runner also produces debug
information in an adhoc manner multiple times in different places, which
really sucks.

This PR ports over the
[frktest](https://github.com/itsfrank/frktest/blob/main/src/assert_impl.luau)
implementation of assertions. Rather than having `assertions`
immediately yield control flow via `error`, all assertions internally
return a typed `failure?`. When this failure is emitted, we extract all
of the debug information needed for tests, exactly once at the point of
the assertion 'failure'. This means that the test runner no longer has
to know how to produce debug information (woo separation of concerns!),
and the actual assertion implementation is responsible for this.

Another freebie is that we only have to produce stacktraces in exactly
one place - I've gone ahead and added that, along with updating the
runtime error tests.

## Future Work
- The stacktraces currently produces are not very good - this is because
`lute run` compiles the script with all optimizations on, and the inling
negatively impacts the quality of stacktraces. I discussed an option
with @vrn-sn to lower the optimization level for .test.luau files.
- The current implementation of `asserts` behaves like `REQUIRE` -
control flow is immediately stopped at this point.
Since all assertion implementations are wrapped by a function that
determines how to handle the control flow, we can now also provide an
implementation of `CHECK`, where you can fail a test and resume control
flow.